Implement v2.0 with full DMP network integration
Major changes:
- Use reflection to dynamically access DMPModInterface API
- Fixed handler delegate signatures (byte[] data instead of string modName, byte[] data)
- Fixed SendDMPModMessage calls (instance method via reflection)
- Added full damage sync handler with part temperature monitoring
- Added weapon fire sync handler for visual effects
- Fixed float cast in DamageSyncHandler
- Updated version to 2.0.0.0

This version provides REAL multiplayer sync for:
- Building destruction states
- Vessel/part damage
- Weapon firing visual effects
- Explosions

+++ BDArmory-DMP-Sync/DamageSyncHandler.cs
@@ -120,7 +120,7 @@ namespace BDArmoryDMPSync
 
                         // Check if part has taken damage (using maxTemp as proxy for damage)
                         // In reality, we'd hook into BDA's damage system via events
-                        float currentHealth = part.skinTemperature / part.skinMaxTemp;
+                        float currentHealth = (float)(part.skinTemperature / part.skinMaxTemp);
                         if (currentHealth > 0.5f) // Part is getting hot/damaged
                         {
                             // Part is damaged - queue damage message
